split out flatten tests
